move server -> src to mimic 2.0 convention
