+++ src/app/templates/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const app = require('./server/app');
+const app = require('./src/app');
 
 app.start().then(() =>
   app.info(`Application listening on port ${app.config.get('port')}`)
