+++ src/app/templates/process.yml
@@ -4,4 +4,4 @@ apps:
     instances: 1
     exec_mode: fork
     no_daemon: true
-    watch: ["server"]
+    watch: ["src"]
