+++ src/app/templates/src/middleware/cors.js
