+++ src/app/templates/src/middleware/favicon.js
