+++ src/app/templates/src/middleware/logger.js
