+++ src/app/templates/src/middleware/static.js
