+++ src/hook/generator.js
@@ -16,8 +16,8 @@ import { hooks as mount } from '../utils/mount';
 module.exports = function (prompt, done, options) {
   const metalsmith = Metalsmith(TEMPLATE_PATH);
   const SERVICE_PATH = path.resolve(options.path);
-  const FEATHERS_PATH = 'server/feathers.json';
-  const MOUNT_PATH = options.mount || 'server/feathers.json';
+  const FEATHERS_PATH = 'src/feathers.json';
+  const MOUNT_PATH = options.mount || 'src/feathers.json';
   const CONFIG_PATH = options.config || 'config';
 
   // if generating as standalone, use src dir
