+++ src/service/generator.js
@@ -16,7 +16,7 @@ import { services as mount } from '../utils/mount';
 
 module.exports = function (prompt, done, options) {
   const metalsmith = Metalsmith(TEMPLATE_PATH);
-  const MOUNT_PATH = options.mount || 'server/feathers.json';
+  const MOUNT_PATH = options.mount || 'src/feathers.json';
   const CONFIG_PATH = options.config || 'config';
 
   // if generating as standalone, use src dir
